requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.0.1
click==8.1.8
ecdsa==0.19.1
fastapi==0.115.12
greenlet==3.2.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pyasn1==0.4.8
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
python-dotenv==1.1.0
python-jose==3.4.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
